+++ .files/modules/oh-my-zsh/install.z.sh
@@ -1,18 +1,22 @@
 # =========================================================================== [Installation] ===== #
 
 # Install oh-my-zsh
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+if [ ! -d "$HOME/.oh-my-zsh" ]; then
+  sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+fi
 
 # Install Antigen
 if [[ "$OS_TYPE" == "stm" ]]; then
-  nix-env -iA nixpkgs.antigen
+  [ -f "$HOME/.nix-profile/share/antigen/antigen.zsh" ] || nix-env -iA nixpkgs.antigen
 fi
 
 if [[ "$OS_TYPE" == "mac" ]]; then
-  brew install antigen
+  [ -f "$(brew --prefix)/share/antigen/antigen.zsh" ] || brew install antigen
 fi
 
 if [[ "$OS_TYPE" == "win" ]]; then
-  sudo apt-get update
-  sudo apt-get install -y zsh-antigen
+  if [ ! -f /usr/share/zsh-antigen/antigen.zsh ]; then
+    sudo apt-get update
+    sudo apt-get install -y zsh-antigen
+  fi
 fi
